refactor: backend
move fgb-to-geojson helper endpoint to helper routes
